split off code into quaternion_simd.rs
