+++ src/lib.rs
@@ -99,6 +99,10 @@ mod structure;
 
 mod matrix;
 mod quaternion;
+
+#[cfg(feature = "simd")]
+mod quaternion_simd;
+
 mod vector;
 
 #[cfg(feature = "simd")]
